Refactor CLAP extensions and helper UI modules
